susoft_service/ocr_service/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR Engines
paddleocr==2.7.3
paddlepaddle==2.6.1

# Image Processing
numpy>=1.26.0,<2.0.0
Pillow==10.1.0
opencv-python-headless>=4.8.0,<4.10.0

# HTTP Client
httpx==0.25.1
aiofiles==23.2.1

# Cache & DB
redis==5.0.1
hiredis==2.2.3

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-jose[cryptography]==3.3.0

google-cloud-vision==3.4.5
